Conflicts:
	src/main/java/com/emc/mongoose/run/Main.java
	src/main/java/com/emc/mongoose/web/ui/StartServlet.java
	src/test/resources/py/chain.py

+++ src/main/java/com/emc/mongoose/util/persist/LevelEntity.java
@@ -26,7 +26,7 @@ implements Serializable{
 	@Id
 	@GeneratedValue(strategy = IDENTITY)
 	@Column(name = "id")
-	private BigInteger id;
+	private long id;
 	@Column(name = "name")
 	private String name;
 	@OneToMany(targetEntity=MessageEntity.class, fetch = FetchType.LAZY, mappedBy = "level")
@@ -44,10 +44,10 @@ implements Serializable{
 	public final void setName(final String name) {
 		this.name = name;
 	}
-	public final BigInteger getId() {
+	public final long getId() {
 		return id;
 	}
-	public final void setId(final BigInteger id) {
+	public final void setId(final long id) {
 		this.id = id;
 	}
 	public final Set<MessageEntity> getMessageSet() {
